move fixtures to standalone resources
