+++ skills/tests/test_export/resources/IntegerMultiplyBy2UDF.scala
@@ -1,0 +1,7 @@
+package com.udf
+
+class IntegerMultiplyBy2UDF extends Function1[Integer, Integer] with Serializable {
+  override def apply(value: Integer): Integer = {
+    if (value == null) null else value * 2
+  }
+}
