+++ skills/tests/test_export/resources/integer_multiply_by_2.sql
@@ -1,0 +1,3 @@
+SELECT *,
+  value * 2 AS result
+FROM test_table
